refactored apilogin
