+++ internal/pkg/interfaces/semsdata.go
